Lab01 commit

+++ src/main/java/inventory/controller/MainScreenController.java
@@ -15,6 +15,8 @@ import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
 import java.net.URL;
@@ -22,12 +24,11 @@ import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class MainScreenController implements Initializable,Controller {
+    private static final Logger logger = LogManager.getLogger(MainScreenController.class);
 
     // Declare fields
-    private Stage stage;
-    private Parent scene;
-    private static Part modifyPart;
-    private static Product modifyProduct;
+
+
     private static int modifyPartIndex;
     private static int modifyProductIndex;
 
@@ -79,8 +80,6 @@ public class MainScreenController implements Initializable,Controller {
     @FXML
     private TextField productsSearchTxt;
 
-    public MainScreenController(){}
-
     public void setService(InventoryService service){
         this.service=service;
         partsTableView.setItems(service.getAllParts());
@@ -113,9 +112,10 @@ public class MainScreenController implements Initializable,Controller {
      * @throws IOException
      */
     private void displayScene(ActionEvent event, String source) throws IOException {
+        Stage stage;
+        Parent scene;
         stage = (Stage)((Button)event.getSource()).getScene().getWindow();
         FXMLLoader loader= new FXMLLoader(getClass().getResource(source));
-        //scene = FXMLLoader.load(getClass().getResource(source));
         scene = loader.load();
         Controller ctrl=loader.getController();
         ctrl.setService(service);
@@ -138,11 +138,13 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
-            System.out.println("Part deleted.");
-            service.deletePart(part);
-        } else {
-            System.out.println("Canceled part deletion.");
+        if(result.isPresent()) {
+            if (result.get() == ButtonType.OK) {
+                logger.info("Part deleted.");
+                service.deletePart(part);
+            } else {
+                logger.info("Canceled part deletion.");
+            }
         }
     }
 
@@ -161,11 +163,13 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete product " + product.getName() + " from products?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
-            service.deleteProduct(product);
-            System.out.println("Product " + product.getName() + " was removed.");
-        } else {
-            System.out.println("Product " + product.getName() + " was not removed.");
+        if(result.isPresent()){
+            if (result.get() == ButtonType.OK) {
+                service.deleteProduct(product);
+                logger.info("Product " + product.getName() + " was removed.");
+            } else {
+                logger.info("Product " + product.getName() + " was not removed.");
+            }
         }
     }
 
@@ -197,6 +201,7 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyPart(ActionEvent event) throws IOException {
+        Part modifyPart;
         modifyPart = partsTableView.getSelectionModel().getSelectedItem();
         modifyPartIndex = service.getAllParts().indexOf(modifyPart);
 
@@ -210,6 +215,7 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyProduct(ActionEvent event) throws IOException {
+        Product modifyProduct;
         modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
         modifyProductIndex = service.getAllProducts().indexOf(modifyProduct);
 
@@ -228,11 +234,13 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Exit");
         alert.setContentText("Are you sure you want to exit?");
         Optional<ButtonType> result = alert.showAndWait();
-        if(result.get() == ButtonType.OK) {
-            System.out.println("Ok selected. Program exited");
-            System.exit(0);
-        } else {
-            System.out.println("Cancel clicked.");
+        if(result.isPresent()){
+            if(result.get() == ButtonType.OK) {
+                logger.info("Ok selected. Program exited");
+                System.exit(0);
+            } else {
+                logger.info("User canceled exit.");
+            }
         }
     }
